mcp-server: make the projection internal and put each failure on its namespace
Nothing outside the package imported `Projection` — it was the same reflex as the others,
so it joins them under internal/. `GatewayError` becomes `Gateway.Error` with a
`Gateway.error(cause)` constructor, which retires the `errorMessage` helper from the API
(hosts were each unwrapping causes by hand), and the tool-call failure becomes
`Server.ToolFailure` / `Server.failure`.

The failure type itself sits in internal/ so the space helpers can use it without a cycle
back through `Server`.

The API is now two namespaces: implement a `Gateway`, build a `Server`.
