+++ packages/core/compute/mcp-server/src/index.ts
@@ -3,6 +3,4 @@
 //
 
 export * as Gateway from './Gateway';
-export * as Projection from './Projection';
 export * as Server from './Server';
-export * from './errors';
