+++ packages/core/compute/mcp-server/src/internal/errors.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export const errorMessage = (error: unknown): string =>
+  error instanceof Error ? (error.message ?? String(error)) : String(error);
