+++ packages/core/compute/mcp-server/src/internal/space.ts
@@ -7,7 +7,7 @@ import * as Schema from 'effect/Schema';
 
 import { SpaceId } from '@dxos/keys';
 
-import { type ToolFailure, failure } from '../errors';
+import { type ToolFailure, failure } from './failure';
 
 export const idParameter = Schema.optional(Schema.String).annotate({
   // The default exists so a read is cheap to issue, not so a write can skip choosing. It is
